+++ src/components/notifications/MessageBox.tsx
@@ -4,7 +4,7 @@ import TimeIndicator from './TimeIndicator';
 import { useNavigate } from 'react-router-dom';
 import { FC, useState } from 'react';
 import { InterfaceBoxProps } from 'native-base/lib/typescript/components/primitives/Box';
-import LeavePageModal from './LeavePageModal';
+import LeavePageModal from '../../modals/LeavePageModal';
 import { Concrete_Notification } from '../../gql/graphql';
 import AppointmentCancelledModal from './NotificationModal';
