+++ src/pages/pupil/single-course/PupilCourseButtons.tsx
@@ -9,7 +9,7 @@ import { getTrafficStatus } from '../../../Utility';
 import WaitinglistBanner from '../../../widgets/WaitinglistBanner';
 import AlertMessage from '../../../widgets/AlertMessage';
 import OpenCourseChatButton from '../../subcourse/OpenCourseChatButton';
-import { canJoinMeeting } from '../../../widgets/appointment/AppointmentDay';
+import { canJoinMeeting } from '../../../widgets/AppointmentDay';
 import { DateTime } from 'luxon';
 import { gql } from '../../../gql';
 import VideoButton from '../../../components/VideoButton';
