+++ src/pages/student/single-course/StudentCourseButtons.tsx
@@ -4,21 +4,11 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { Lecture, Subcourse } from '../../../gql/graphql';
 import { useLayoutHelper } from '../../../hooks/useLayoutHelper';
-import { canJoinMeeting } from '../../../widgets/appointment/AppointmentDay';
+import { canJoinMeeting } from '../../../widgets/AppointmentDay';
 import { DateTime } from 'luxon';
 import { useMemo } from 'react';
 import OpenCourseChatButton from '../../subcourse/OpenCourseChatButton';
-import ContactParticipants from './ContactParticipants';
 import VideoButton from '../../../components/VideoButton';
-
-type SubcourseOfStudent = {
-    id: number;
-    published: boolean;
-    isInstructor: boolean;
-    canCancel: { allowed: boolean };
-    canContactParticipants: { allowed: boolean };
-    canEdit: { allowed: boolean };
-};
 
 type ActionButtonProps = {
     subcourse: Pick<
