+++ src/widgets/CourseBanner.tsx
